Gets this code to compile and run.
See the instruction in compile-me-hack.include.cpp for how to get the
code to compile and run.  The added file "fixes" the issue with the
missing definition of Utils::Output by abusing the C preprocessor.

Assuming that Utils::Output prints its' arguments to the screen, this
will dump information to the log file for inspection.